chore: Cleanup native tests
