+++ presto-native-tests/src/test/java/com/facebook/presto/nativetests/TestCustomFunctions.java
@@ -15,6 +15,7 @@ package com.facebook.presto.nativetests;
 
 import com.facebook.presto.testing.QueryRunner;
 import com.facebook.presto.tests.AbstractTestQueryFramework;
+import com.google.common.collect.ImmutableList;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.BeforeSuite;
 import org.testng.annotations.Test;
@@ -27,8 +28,11 @@ import java.io.InputStreamReader;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.util.List;
+import java.util.Optional;
 
-import static com.facebook.presto.nativetests.NativeTestsUtils.getCustomFunctionsPluginDirectory;
+import static com.facebook.presto.nativeworker.PrestoNativeQueryRunnerUtils.nativeHiveQueryRunnerBuilder;
+import static com.facebook.presto.sidecar.NativeSidecarPluginQueryRunnerUtils.setupNativeSidecarPlugin;
 import static java.lang.Boolean.parseBoolean;
 
 public class TestCustomFunctions
@@ -87,13 +91,17 @@ public class TestCustomFunctions
     @Override
     protected QueryRunner createQueryRunner() throws Exception
     {
-        return NativeTestsUtils.createNativeQueryRunner(storageFormat, false, sidecarEnabled, true);
-    }
-
-    @Override
-    protected void createTables()
-    {
-        NativeTestsUtils.createTables(storageFormat);
+        QueryRunner queryRunner = nativeHiveQueryRunnerBuilder()
+                .setStorageFormat(storageFormat)
+                .setAddStorageFormatToPath(true)
+                .setUseThrift(true)
+                .setCoordinatorSidecarEnabled(sidecarEnabled)
+                .setPluginDirectory(Optional.of(getCustomFunctionsPluginDirectory().toString()))
+                .build();
+        if (sidecarEnabled) {
+            setupNativeSidecarPlugin(queryRunner);
+        }
+        return queryRunner;
     }
 
     /// Sidecar is needed to support custom functions in Presto C++.
@@ -144,5 +152,34 @@ public class TestCustomFunctions
                     "SELECT map_size(MAP(ARRAY[1,2], ARRAY[10,20]))",
                     "line 1:8: Function map_size not registered");
         }
+    }
+
+    private static Path getCustomFunctionsPluginDirectory()
+            throws Exception
+    {
+        Path prestoRoot = findPrestoRoot();
+
+        // All candidate paths relative to prestoRoot
+        List<Path> candidates = ImmutableList.of(
+                prestoRoot.resolve("presto-native-tests/_build/debug/presto_cpp/tests/custom_functions"),
+                prestoRoot.resolve("presto-native-tests/_build/release/presto_cpp/tests/custom_functions"));
+
+        return candidates.stream()
+                .filter(Files::exists)
+                .findFirst()
+                .orElseThrow(() -> new IllegalStateException("Could not locate custom functions directory"));
+    }
+
+    private static Path findPrestoRoot()
+    {
+        Path dir = Paths.get(System.getProperty("user.dir")).toAbsolutePath();
+        while (dir != null) {
+            if (Files.exists(dir.resolve("presto-native-tests")) ||
+                    Files.exists(dir.resolve("presto-native-execution"))) {
+                return dir;
+            }
+            dir = dir.getParent();
+        }
+        throw new IllegalStateException("Could not locate presto root directory.");
     }
 }
